% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_footprints.R
\name{circularize}
\alias{circularize}
\title{Apply 5' recovery bias to footprints}
\usage{
circularize(footprints, circ_bias)
}
\arguments{
\item{footprints}{list of \code{footprint} objects}

\item{circ_bias}{named numeric vector; probabilities for successful 5' circularization}
}
\value{
subset of input \code{footprints} that were successfully "circularized"
}
\description{
Apply 5' recovery bias to footprints
}
