% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{count_codons}
\alias{count_codons}
\title{Compute codon proportions in transcriptome}
\usage{
count_codons(transcript_seq, codons, utr5_length, utr3_length)
}
\arguments{
\item{transcript_seq}{list of character vectors; output from read_fasta()}

\item{codons}{character vector; codons in transcriptome}

\item{utr5_length}{integer; length of 5' UTR}

\item{utr3_length}{integer; length of 3' UTR}
}
\value{
matrix of codon appearances per transcript
}
\description{
Compute codon proportions in transcriptome
}
