% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_footprints.R
\name{digest}
\alias{digest}
\title{Generate RPF sequences for all transcripts}
\usage{
digest(
  transcript_seq,
  rpf_counts,
  delta_5,
  delta_3,
  min_size = 27,
  max_size = 31,
  digest_transcript
)
}
\arguments{
\item{transcript_seq}{list of character vectors; output from read_fasta() or simulate_transcriptome()}

\item{rpf_counts}{list of numeric vectors; output from simulate_profiles()}

\item{delta_5}{named numeric_vector; probabilities of 5' digest lengths}

\item{delta_3}{named numeric vector; probabilities of 3' digest lengths}

\item{min_size}{integer; minimum RPF size}

\item{max_size}{integer; maximum RPF size}

\item{digest_transcript}{function; function that digests individual transcripts}

\item{mc.cores}{integer; number of cores to use for parallelization}
}
\value{
list of \code{footprint} objects
}
\description{
Generate RPF sequences for all transcripts
}
