% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transcriptome.R
\name{generate_transcript}
\alias{generate_transcript}
\title{Generate simulated mRNA transcript (5' UTR + CDS + 3' UTR)}
\usage{
generate_transcript(num_codons, codon_distribution)
}
\arguments{
\item{num_codons}{integer; length of  transcript in codons}

\item{codon_distribution}{named numeric vector; codon proportions in transcriptome}
}
\value{
character vector of \code{length(num_codons)} of mRNA transcript in trinucleotides
}
\description{
Generate simulated mRNA transcript (5' UTR + CDS + 3' UTR)
}
