% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_footprints.R
\name{get_bias_region}
\alias{get_bias_region}
\title{Get sequence for nucleotides at RPF edges}
\usage{
get_bias_region(footprints, region, bias_length)
}
\arguments{
\item{footprints}{list of \code{footprint} objects}

\item{region}{character; "f5" for 5' region or "f3" for 3' region}

\item{bias_length}{integer; number of nucleotides to constitute end region}
}
\value{
character vector of sequences corresponding to bias region
}
\description{
Get sequence for nucleotides at RPF edges
}
