% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_footprints.R
\name{ligate}
\alias{ligate}
\title{Apply 3' recovery bias to footprints}
\usage{
ligate(footprints, lig_bias)
}
\arguments{
\item{footprints}{list of \code{footprint} objects}

\item{lig_bias}{named numeric vector; probabilities for successful 3' ligation}
}
\value{
subset of input \code{footprints} that were successfully "ligated"
}
\description{
Apply 3' recovery bias to footprints
}
