% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_cts_by_codon}
\alias{read_cts_by_codon}
\title{Read in cts_by_codon file from iXnos}
\usage{
read_cts_by_codon(input_file)
}
\arguments{
\item{input_file}{character; path to input file}
}
\value{
list of numeric vectors; ribosome occupancies by codon position
}
\description{
Read in cts_by_codon file from iXnos
}
