% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_fasta}
\alias{read_fasta}
\title{Read in transcriptome fasta file, convert into vector of codons}
\usage{
read_fasta(fasta_file, utr5_length, utr3_length)
}
\arguments{
\item{fasta_file}{character; path to transcriptome fasta file}

\item{utr5_length}{integer; length of 5' UTR}

\item{utr3_length}{integer; length of 3' UTR}
}
\value{
list of character vectors of mRNA transcripts in trinucleotides
}
\description{
Read in transcriptome fasta file, convert into vector of codons
}
