% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_raw_profiles}
\alias{read_raw_profiles}
\title{Count codon occurrences by transcript}
\usage{
read_raw_profiles(input_file)
}
\arguments{
\item{input_file}{character; path to input file}
}
\value{
list of numeric vectors; ribosome counts per codon per transcript
}
\description{
Count codon occurrences by transcript
}
