% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_footprints.R
\name{RT}
\alias{RT}
\title{Add extra non-templated base at 5' end of footprints}
\usage{
RT(footprints, rt_bias)
}
\arguments{
\item{footprints}{list of \code{footprint} objects}

\item{rt_bias}{named numeric vector; probabilities of adding additional nt}
}
\value{
input \code{footprints} with some footprints added NTA base
}
\description{
Add extra non-templated base at 5' end of footprints
}
