% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ribosome_distribution.R
\name{simulate_pi}
\alias{simulate_pi}
\title{Generate per-codon-position probabilities}
\usage{
simulate_pi(transcript_seq, utr5_length = 6, utr3_length = 4, codon_TE)
}
\arguments{
\item{transcript_seq}{list of character vectors; output from read_fasta() or simulate_transcriptome()}

\item{utr5_length}{integer; length of 5' UTR in "codons"}

\item{utr3_length}{integer; length of 3' UTR in "codons}

\item{codon_TE}{named numeric vector; TE of codon averaged over transcriptome}
}
\description{
Generate per-codon-position probabilities
}
