% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_footprints.R
\name{simulate_profiles}
\alias{simulate_profiles}
\title{Generate simulated profile of ribosome occupancy by codon position per transcript}
\usage{
simulate_profiles(nRibosomes, rhos, pis)
}
\arguments{
\item{rhos}{numeric vector; multinomial probabilities for ribosomes per transcript; output from simulate_rho()}

\item{pis}{list of numeric vectors; multinomial probabilities for ribosomes per codon position; output from simulate_pi()}

\item{num_ribosomes}{integer; number of footprints to simulate}
}
\value{
list of numeric vector of ribosome occupancies by codon position
}
\description{
Generate simulated profile of ribosome occupancy by codon position per transcript
}
