% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ribosome_distribution.R
\name{simulate_rho}
\alias{simulate_rho}
\title{Generate per-transcript probabilities from loess model trained on experimental data}
\usage{
simulate_rho(transcript_seq, expt_lengths, expt_abundances)
}
\arguments{
\item{transcript_seq}{list of character vectors; output from read_fasta() or simulate_transcriptome()}

\item{expt_lengths}{numeric vector; transcript lengths from a model experiment}

\item{expt_abundances}{numeric vector; transcript abundances from a model experiment}
}
\value{
numeric vector of per-transcript probabilities
}
\description{
Generate per-transcript probabilities from loess model trained on experimental data
}
