% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transcriptome.R
\name{simulate_transcriptome}
\alias{simulate_transcriptome}
\title{Generate a simulated transcriptome}
\usage{
simulate_transcriptome(transcript_lengths, codon_distribution)
}
\arguments{
\item{transcript_lengths}{numeric vector; transcript lengths in codons}

\item{codon_distribution}{named numeric vector; codon proportions in transcriptome}
}
\value{
list of \code{length(transcript_lengths)} of character vectors corresponding to transcript sequences
}
\description{
Generate a simulated transcriptome
}
