% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{write_footprints_fasta}
\alias{write_footprints_fasta}
\title{Write footprints to .fasta file}
\usage{
write_footprints_fasta(footprints, output_file)
}
\arguments{
\item{footprints}{list of \code{footprint} objects}

\item{output_file}{character; path to output .fasta file}
}
\description{
Write footprints to .fasta file
}
