% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{write_footprints_fastq}
\alias{write_footprints_fastq}
\title{Write footprints to .fastq file}
\usage{
write_footprints_fastq(footprints, output_file, adapterr = "CTGTAGGCACCATCAAT")
}
\arguments{
\item{footprints}{list of \code{footprint} objects}

\item{output_file}{character; path to output .fastq file}

\item{adapterr}{character; 3' adapter to ligate onto footprint sequence}
}
\description{
Write footprints to .fastq file
}
