% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{write_transcriptome}
\alias{write_transcriptome}
\title{Write output from simulate_transcriptome() to .fasta file}
\usage{
write_transcriptome(transcripts, output_file)
}
\arguments{
\item{transcripts}{list of character vectors; transcript sequences from simTranscriptome()}

\item{output_file}{character; path to output .fasta file}
}
\description{
Write output from simulate_transcriptome() to .fasta file
}
